core/requirements.txt
Django==3.2.4
django-datatables-view==1.19.1
django-extensions==3.0.9
django-filter
django-widget-tweaks
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.17.1
Jinja2==3.0.1
Pillow
PyJWT==2.0.1
requests==2.25.1
Unidecode==1.2.0
urllib3==1.26.5
python-dotenv==0.19.2
humanize
django-cors-headers
django-resized
django_compressor==4.4
django-pwa
django-celery-beat
gunicorn
schedule
psycopg2-binary
whitenoise
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-apps-meet
pytest
pytest-django
selenium
stripe==5.5.0
django-minio-storage
python-decouple